requirements.txt
feedparser==6.0.11
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
msal==1.28.0
PyYAML==6.0.2
pdfminer.six==20240706
python-dotenv==1.0.1
